+++ sdk/dictKey/dictKey.js
@@ -20,19 +20,14 @@ class DictKey extends Resource {
      * @param ownerType [string, default null]: Pix key owner type. ex 'naturalPerson' or 'legalPerson'
      * @param bankName [string, null null]: bank name associated with the DICT key. ex: 'Stark Bank'
      * @param ispb [string, default null]: ISPB code used for transactions. ex: '20018183'
-     * @param branchCode [string, default null]: bank account branch code associated with the Pix key. ex: '9585'
-     * @param accountNumber [string, default null]: bank account number associated with the Pix key. ex: '9828282578010513'
-     * @param accountType [string, default null]: bank account type associated with the Pix key. ex: 'checking', 'saving', 'salary' or 'payment'
+     * @param branchCode [string, default null]: encrypted bank account branch code associated with the Pix key. ex: 'ZW5jcnlwdGVkLWJyYW5jaC1jb2Rl'
+     * @param accountNumber [string, default null]: encrypted bank account number associated with the Pix key. ex: 'ZW5jcnlwdGVkLWFjY291bnQtbnVtYmVy'
+     * @param accountType [string, default null]: bank account type associated with the Pix key. ex: 'checking', 'savings', 'salary' or 'payment'
      * @param status [string, default null]: current Pix key status. ex: 'created', 'registered', 'canceled' or 'failed'
-     * @param accountCreated [string, default null]: creation datetime of the bank account associated with the Pix key. ex: '2020-11-05T14:55:08.812665+00:00'
-     * @param owned [string, default null]: datetime since when the current owner hold this Pix key. ex : '2020-11-05T14:55:08.812665+00:00'     
-     * @param created [string, default null]: creation datetime for the Pix key. ex: '2020-11-05T14:55:08.812665+00:00'
-     *
      */
     constructor({
-                    id, type = null, accountCreated = null, accountType = null, name = null, taxId = null,
-                    ownerType = null, bankName = null, ispb = null, branchCode = null, accountNumber = null, 
-                    status = null, owned = null, created = null
+                    id, type = null, accountCreated = null, accountType = null, name = null, taxId = null, ownerType = null,
+                    bankName = null, ispb = null, branchCode = null, accountNumber = null, status = null 
                 }) {
         super(id);
         this.type = type;
@@ -45,9 +40,7 @@ class DictKey extends Resource {
         this.ispb = ispb;
         this.branchCode = branchCode;
         this.accountNumber = accountNumber;
-        this.status = status;
-        this.owned = check.date(owned);
-        this.created = check.date(created);
+        this.status = status
     }
 }